splitting up TypeUtils
