+++ src/type/typeUtils/index.ts
@@ -1,0 +1,8 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+export * from "./equalType";
+export * from "./factories";
+export * from "./isSubset";
+export * from "./primitive";
+export * from "./typeUtils";
